Added the neural network to the snake.
Every snake object will now have a neural network instance in it. The
network will be initialized to a random value at start.

Need to add the genetic algorithm for it to work, but no real signs of
progress. Even after running it many times all the snakes would just go
with their initial direction until they hit a wall.